+++ System.IO.Abstractions.TestingHelpers/MockFileStream.cs
@@ -8,11 +8,7 @@
         private readonly bool canWrite = true;
         private readonly FileOptions options;
 
-#if NET40
-        private bool closed;
-#else
         private bool disposed;
-#endif
 
         public enum StreamType
         {
@@ -39,12 +35,12 @@
             this.mockFileDataAccessor = mockFileDataAccessor ?? throw new ArgumentNullException(nameof(mockFileDataAccessor));
             this.path = path;
             this.options = options;
-            
+
             if (mockFileDataAccessor.FileExists(path))
             {
                 var fileData = mockFileDataAccessor.GetFile(path);
                 fileData.CheckFileAccess(path, streamType != StreamType.READ ? FileAccess.Write : FileAccess.Read);
-                
+
                 /* only way to make an expandable MemoryStream that starts with a particular content */
                 var data = fileData.Contents;
                 if (data != null && data.Length > 0 && streamType != StreamType.TRUNCATE)
@@ -69,25 +65,12 @@
 
                 mockFileDataAccessor.AddFile(path, new MockFileData(new byte[] { }));
             }
-            
+
             canWrite = streamType != StreamType.READ;
         }
 
         public override bool CanWrite => canWrite;
 
-#if NET40
-        public override void Close()
-        {
-            if (closed)
-            {
-                return;
-            }
-            InternalFlush();
-            base.Close();
-            OnClose();
-            closed = true;
-        }
-#else
         protected override void Dispose(bool disposing)
         {
             if (disposed)
@@ -99,7 +82,6 @@
             OnClose();
             disposed = true;
         }
-#endif
 
         public override void Flush()
         {
@@ -128,12 +110,10 @@
                 mockFileDataAccessor.RemoveFile(path);
             }
 
-#if NET40
             if (options.HasFlag(FileOptions.Encrypted) && mockFileDataAccessor.FileExists(path))
             {
                 mockFileDataAccessor.FileInfo.FromFileName(path).Encrypt();
             }
-#endif
         }
     }
 }
