+++ System.IO.Abstractions.TestingHelpers/MockFileStreamFactory.cs
@@ -29,15 +29,12 @@ namespace System.IO.Abstractions.TestingHelpers
         public Stream Create(string path, FileMode mode, FileAccess access, FileShare share, int bufferSize, bool useAsync)
             => new MockFileStream(mockFileSystem, path, GetStreamType(mode, access));
 
-#if NET40
         public Stream Create(string path, FileMode mode, FileSystemRights rights, FileShare share, int bufferSize, FileOptions options, FileSecurity fileSecurity)
             => new MockFileStream(mockFileSystem, path, GetStreamType(mode), options);
 
         public Stream Create(string path, FileMode mode, FileSystemRights rights, FileShare share, int bufferSize, FileOptions options)
             => new MockFileStream(mockFileSystem, path, GetStreamType(mode), options);
-#endif
 
-#if NET40 || NETSTANDARD_20
         [Obsolete("This method has been deprecated. Please use new Create(SafeFileHandle handle, FileAccess access) instead. http://go.microsoft.com/fwlink/?linkid=14202")]
         public Stream Create(IntPtr handle, FileAccess access)
             => new MockFileStream(mockFileSystem, handle.ToString(), GetStreamType(FileMode.Append, access));
@@ -53,7 +50,6 @@ namespace System.IO.Abstractions.TestingHelpers
         [Obsolete("This method has been deprecated. Please use new Create(SafeFileHandle handle, FileAccess access, int bufferSize, bool isAsync) instead, and optionally make a new SafeFileHandle with ownsHandle=false if needed. http://go.microsoft.com/fwlink/?linkid=14202")]
         public Stream Create(IntPtr handle, FileAccess access, bool ownsHandle, int bufferSize, bool isAsync)
             => new MockFileStream(mockFileSystem, handle.ToString(), GetStreamType(FileMode.Append, access));
-#endif
 
         public Stream Create(SafeFileHandle handle, FileAccess access)
             => new MockFileStream(mockFileSystem, handle.ToString(), GetStreamType(FileMode.Append, access));
@@ -70,7 +66,7 @@ namespace System.IO.Abstractions.TestingHelpers
             {
                 return MockFileStream.StreamType.READ;
             }
-            else if (mode == FileMode.Append) 
+            else if (mode == FileMode.Append)
             {
                 return MockFileStream.StreamType.APPEND;
             }
