+++ System.IO.Abstractions/IDirectory.cs
@@ -12,7 +12,8 @@ namespace System.IO.Abstractions
 
         /// <inheritdoc cref="Directory.CreateDirectory(string)"/>
         IDirectoryInfo CreateDirectory(string path);
-
+        /// <inheritdoc cref="Directory.CreateDirectory(string,DirectorySecurity)"/>
+        IDirectoryInfo CreateDirectory(string path, DirectorySecurity directorySecurity);
         /// <inheritdoc cref="Directory.Delete(string)"/>
         void Delete(string path);
         /// <inheritdoc cref="Directory.Delete(string,bool)"/>
